slider

+++ PocketPlanets/PocketPlanets.pde
@@ -25,8 +25,7 @@ void draw()
 
 class State
 {
-  void update() {
-  }
+  void update() {}
 }
 
 class Button
@@ -67,9 +66,9 @@ class Button
 
 class Slider
 {
-  int x, y, w, min, max; 
+  float x, y, w, min, max; 
   float current;
-  int sx;
+  float sx;
   int colour;
 
   void displaySlider()
@@ -113,7 +112,7 @@ class Slider
     }
   }
 
-  Slider(int a, int b, int c, int d, int e, int f)
+  Slider(float a, float b, float c, float d, float e, int f)
   {
     x = a;
     sx = a;
